refactor: restores Negotiation Path
