+++ src/main/java/eu/bbmri_eric/negotiator/configuration/state_machine/negotiation/InitializeStateForResourceAction.java
@@ -1,7 +1,7 @@
 package eu.bbmri_eric.negotiator.configuration.state_machine.negotiation;
 
 import eu.bbmri_eric.negotiator.configuration.state_machine.resource.NegotiationResourceState;
-import eu.bbmri_eric.negotiator.database.model.negotiation.Negotiation;
+import eu.bbmri_eric.negotiator.database.model.Negotiation;
 import eu.bbmri_eric.negotiator.database.repository.NegotiationRepository;
 import eu.bbmri_eric.negotiator.service.UserNotificationService;
 import jakarta.transaction.Transactional;
