+++ src/main/java/eu/bbmri_eric/negotiator/database/model/Negotiation.java
@@ -1,14 +1,8 @@
-package eu.bbmri_eric.negotiator.database.model.negotiation;
+package eu.bbmri_eric.negotiator.database.model;
 
 import com.vladmihalcea.hibernate.type.json.JsonType;
 import eu.bbmri_eric.negotiator.configuration.state_machine.negotiation.NegotiationState;
 import eu.bbmri_eric.negotiator.configuration.state_machine.resource.NegotiationResourceState;
-import eu.bbmri_eric.negotiator.database.model.AuditEntity;
-import eu.bbmri_eric.negotiator.database.model.NegotiationLifecycleRecord;
-import eu.bbmri_eric.negotiator.database.model.NegotiationResourceLifecycleRecord;
-import eu.bbmri_eric.negotiator.database.model.PersonNegotiationRole;
-import eu.bbmri_eric.negotiator.database.model.Request;
-import eu.bbmri_eric.negotiator.database.model.Resource;
 import eu.bbmri_eric.negotiator.database.model.attachment.Attachment;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.CollectionTable;
