+++ src/main/java/eu/bbmri_eric/negotiator/database/model/attachment/BaseAttachmentProjection.java
@@ -1,8 +1,8 @@
 package eu.bbmri_eric.negotiator.database.model.attachment;
 
+import eu.bbmri_eric.negotiator.database.model.Negotiation;
 import eu.bbmri_eric.negotiator.database.model.Organization;
 import eu.bbmri_eric.negotiator.database.model.Person;
-import eu.bbmri_eric.negotiator.database.model.negotiation.Negotiation;
 
 public interface BaseAttachmentProjection {
   Person getCreatedBy();
