+++ src/main/java/eu/bbmri_eric/negotiator/service/UserNotificationService.java
@@ -1,8 +1,8 @@
 package eu.bbmri_eric.negotiator.service;
 
+import eu.bbmri_eric.negotiator.database.model.Negotiation;
 import eu.bbmri_eric.negotiator.database.model.Post;
 import eu.bbmri_eric.negotiator.database.model.Resource;
-import eu.bbmri_eric.negotiator.database.model.negotiation.Negotiation;
 import eu.bbmri_eric.negotiator.dto.NotificationDTO;
 import java.util.List;
