+++ src/main/java/eu/bbmri_eric/negotiator/service/UserNotificationServiceImpl.java
@@ -2,7 +2,7 @@ package eu.bbmri_eric.negotiator.service;
 
 import eu.bbmri_eric.negotiator.configuration.state_machine.resource.NegotiationResourceEvent;
 import eu.bbmri_eric.negotiator.database.model.*;
-import eu.bbmri_eric.negotiator.database.model.negotiation.Negotiation;
+import eu.bbmri_eric.negotiator.database.model.Negotiation;
 import eu.bbmri_eric.negotiator.database.repository.NotificationRepository;
 import eu.bbmri_eric.negotiator.database.repository.PersonRepository;
 import eu.bbmri_eric.negotiator.dto.NotificationDTO;
